Refactor: 스타일 정리
