+++ src/components/homePage/search/Search.tsx
@@ -1,0 +1,11 @@
+import * as S from "./Search.styled";
+
+const Search = () => {
+  return (
+    <S.SearchContainer>
+      <input type="text" placeholder="Search" />
+    </S.SearchContainer>
+  );
+};
+
+export default Search;
