+++ src/pages/home/Home.styled.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import { colors } from "../../style/theme/colors";
+
+export const NoMessage = styled.div`
+  margin-top: 40px;
+  text-align: center;
+  color: ${colors.white};
+`;
