+++ src/style/theme/colors.ts
@@ -8,7 +8,21 @@ export const colors = {
 
   gray: "#999",
   gray_900: "#C3C1C1",
-  gray_300: "#D9D9D9",
-  gray_200: "#E4E4E4",
-  gray_100: "#F8F8F8",
+  // gray_300: "#D9D9D9",
+  gray_200: "#c2c2c2",
+  gray_100: "#6f6f6f78",
+
+  gradation_purple: "#8049ff",
+  gradation_pink: "#e641ff",
+
+  pastel_purple: "#b1b1eb",
+
+  message_input: "#545454",
+
+  add_purple: "#8f60ff",
+  add_Pink: "#eb6aff",
+  online_green: "#80fb80",
+
+  modal_inner: "#e7def0",
+  make_chat_pink_btn: "#da9999",
 } as const;
